+++ packages/create-graph/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "./dist/types",
+    "rootDir": "./src",
+    "rewriteRelativeImportExtensions": false
+  },
+  "references": [{ "path": "../graph" }, { "path": "../graph-plugins" }],
+  "include": ["src/**/*.ts", "src/**/*.d.ts"]
+}
